requirements.txt
pyTelegramBotAPI==4.12.0
huggingface_hub==0.27.0
Pillow==9.5.0
python-dotenv==1.0.0

